requirements/local.txt
# -r base.txt # includes the base.txt requirements file

WebTest==2.0.5
coverage==3.6
django-coverage==1.2.2
django-nose==1.1
django-debug-toolbar==0.9.4
django-webtest==1.6.1
django-extensions==1.2.0
factory-boy==2.0.2
ipython==1.0.0
nose==1.3.0
